Don't remove node_modules in CI environment

+++ .ci/scripts/install-armhf-deps.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
-find -type d -name node_modules -print -exec rm -rf '{}' \; || true
+if [ -n "$CI" ]; then
+	find -type d -name node_modules -print -exec rm -rf '{}' \; || true
+fi
 # WARNING: This script needs to be run in sdk-cross:node8-armhf image
 # Install Kuzzle deps using cross build toolchain
 npm install --target_arch=arm  --target_platform=linux --silent --unsafe-perm \
