+++ doc/doc.sh
@@ -20,10 +20,11 @@ fi
 
 case $1 in
   prepare)
-    rm -rf framework
+		if [[ -n "$CI" ]]; then
+			rm -rf framework
+		fi
 
     echo "Clone documentation framework"
-    rm -rf framework/
     git clone --depth 10 --single-branch --branch master https://github.com/kuzzleio/documentation.git framework/
 
     echo "Link local doc for dead links checking"
